Remove skiptutorial. Add /ql for quest lookup due to ID changes. Decayed quest cancellation notice on area reset. Clean up EnemyId/EnemyUIId/ClientItemInfos implementation; use ItemId, EnemyId, and EnemyUIId when possible instead of uints Light quest DB stuff, adjustments to handle variant ids, removing decayed quests when returning to town, rotation task and RPC for child servers. Fix some tasks announcing to the non-head servers on the wrong route. Migration 43 -> 44; add support for new quest_id + variant_index scheme for decoupling quest and schedule id. Cleanup inconsistent packet id naming; add some new quest cancellation packets. Light quest generation, mixin for light quest rewards, and asset. Add DeliveryItemsLight and KillTargetEnemies block support to QuestBlockExtension. Add explicit support for DistributionStart/End in the quest asset and quest system. Remove existing board quests.

+++ Arrowgene.Ddon.GameServer/Quests/Extensions/QuestBlockResultCmdExtension.cs
@@ -238,7 +238,7 @@ namespace Arrowgene.Ddon.GameServer.Quests.Extensions
         {
             questBlock.HandPlayerItems.Add(new QuestItem()
             {
-                ItemId = (uint) itemId,
+                ItemId = itemId,
                 Amount = amount
             });
